attach netrw_priv to smsc95xx_priv
As long as it's been a global object, a relative lock always
required. I don't think it's a good design when this tries to
support more features, not just 'stats'.

It attaches to 'smsc95xx_priv' object as long as the object
is alive, so it doesn't often need to lock/unlock for it.

No longer netrw_priv object as a global object.

Signed-off-by: Leesoo Ahn <lsahn@ooseel.net>